WIP Make compatible with solidus_i18n 2.0
SolidusI18n 2.0 dropped the locales controller. That was moved into
Solidus 2.6

Need to find a way to use the Spree::Admin::LocalesController of
SolidusI18n 1.3 if Solidus version is lower than 2.6 to be able to
update the supported_locales setting. Ideally this would be moved in to
the store settings as well. But this would drop support for SolidusI18n
< 2.0 and Solidus < 2.6