Allow set custom permission to allow/deny translations access
The backend/app/controllers/spree/admin/base_controller.rb uses the
model_class to authorize an admin for a given action.

To allow a user with a custom PermissionSet to access the translations
an `alias_method :model_class, :klass` is enough.

In this way if a PermissionSet say:

class CustomPermission < Spree::PermissionSets::Base
  def activate!
    can :manage, Spree::Product
  end
end

the user will be able to :manage the Spree::Product::Translation too.

Co-authored-by: oeN <oeN@users.noreply.github.com>