Add specs around discarding Spree::ShippingMethod
This is the only model (with Spree::Product) that has
paranoia/discard enabled. This commit does not change the
behavior but adds some specs to describe what's the expected one:
when a shipping method is soft-deleted, translations are kept.

A future improvement could be adding the deleted_at column to
that translation table (as it's done with produts) and adding
discard to that model in order to mark as discarded the
translations associated to a discared shipping method.