Fix bundle issues in test/development
Remove loading a fork when testing against newer
versions of Solidus/Rails.

Now that https://github.com/svenfuchs/routing-filter/pull/76
has been release that's not needed anymore.

We still have to rely on the GitHub version of the gem when
used with Rails 6.1 because the compatibility fix has been
merged but not yet released. see
https://github.com/svenfuchs/routing-filter/issues/75.